+++ TerminalGameEngine/LevelEndFlag.h
@@ -15,7 +15,7 @@ public:
 
 protected:
     virtual bool CanExitScreenSpace() const override { return false; }
-    virtual float GetGravityScale() const override { return 0; }
+    virtual double GetGravityScale() const override { return 0; }
     virtual void OnCollisionEnter(GameObject* other, Direction collisionDirection);
     virtual void OnCollisionExit(Direction collisionDirection){ }
     virtual void InitModel() { SetModel(flagModel); }
